+++ plugins/node/opentelemetry-instrumentation-bunyan/test/bunyan.test.ts
@@ -28,7 +28,7 @@ import {
 import { NodeTracerProvider } from '@opentelemetry/sdk-trace-node';
 import { isWrapped } from '@opentelemetry/instrumentation';
 import { resourceFromAttributes } from '@opentelemetry/resources';
-import { SEMRESATTRS_SERVICE_NAME } from '@opentelemetry/semantic-conventions';
+import { ATTR_SERVICE_NAME } from '@opentelemetry/semantic-conventions';
 import * as assert from 'assert';
 import * as sinon from 'sinon';
 import { Writable } from 'stream';
@@ -47,7 +47,7 @@ tracerProvider.register();
 const tracer = tracerProvider.getTracer('default');
 
 const resource = resourceFromAttributes({
-  [SEMRESATTRS_SERVICE_NAME]: 'test-instrumentation-bunyan',
+  [ATTR_SERVICE_NAME]: 'test-instrumentation-bunyan',
 });
 const loggerProvider = new LoggerProvider({ resource });
 const memExporter = new InMemoryLogRecordExporter();