+++ plugins/node/instrumentation-mongoose/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.49.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-mongoose-v0.48.0...instrumentation-mongoose-v0.49.0) (2025-06-02)
+
+
+### Features
+
+* **deps:** update deps matching '@opentelemetry/*' ([#2871](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2871)) ([d33c6f2](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/d33c6f232a3c5673e618fa62692d2d3bbfe4c0fc))
+* **instrumentation-mongoose:** add instrumentation of static methods ([#2748](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/2748)) ([55cc256](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/55cc2568505179652ce423c5874bccd7404d4b99))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.47.0 to ^0.48.0
+
 ## [0.48.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-mongoose-v0.47.1...instrumentation-mongoose-v0.48.0) (2025-05-15)